+++ src/components/nav-bar/index.tsx
@@ -4,7 +4,6 @@ import { For, Image, Stack, StackProps } from "@chakra-ui/react";
 import {
   faCar,
   faCircleQuestion,
-  faFileShield,
   faFlagCheckered,
   faInfoCircle,
   faLanguage,
@@ -114,14 +113,6 @@ function NavBar({ ...props }: StackProps) {
           >
             <BMCIcon />
           </NavBarButton>
-
-          <NavBarButton
-            key={"pp"}
-            label={"Privacy"}
-            icon={faFileShield}
-            selected={location === ETabs.Privacy}
-            onClick={() => navigate(ETabs.Privacy)}
-          />
 
           <HelpDialog>
             <NavBarButton